+++ client/main.js
@@ -9,25 +9,6 @@ STATUS_PENDING=1;
 STATUS_APPROVED=2;
 STATUS_REJECTED=3;
 
-adminNav = adminNav.concat([
-  {
-    route: 'posts_pending',
-    label: 'Pending'
-  },
-  {
-    route: 'all-users',
-    label: 'Users'
-  },
-  {
-    route: 'settings',
-    label: 'Settings'
-  },
-  {
-    route: 'toolbox',
-    label: 'Toolbox'
-  }   
-]);
-
 // Sort postModules array position using modulePositions as index
 postModules = _.sortBy(postModules, function(module){return _.indexOf(modulePositions, module.position)});